test: update entities and tests for BigInt-only INT64
- Change i32 fields to i64 in test entities (user, product, all_types)
- Update integer literals in type_tests.rs
- Add COUNT + GROUP BY aggregate tests
- Replace .count() with .all().len() due to SeaORM u64 limitation

+++ tests/entity/user.rs
@@ -7,7 +7,7 @@ pub struct Model {
     pub id: String,
     pub name: String,
     pub email: String,
-    pub age: Option<i32>,
+    pub age: Option<i64>,
     pub active: bool,
     pub created_at: DateTime,
 }
